+++ src/views/Admin/components/AdminProduct.vue
@@ -21,7 +21,7 @@
 
       <label class="form-field">
         <span class="form-label">Bilde</span>
-        <img v-if="product.photoURL" :src="product.photoURL" />
+        <img v-if="product.photoURL" :src="product.photoURL" class="preview-image" />
 
         <image-uploader
           :max-width="450"
@@ -39,6 +39,7 @@
       <div class="form-field">
         <label>
           <span class="form-label">Mission statement</span>
+          <span class="form-help">Her kan du skrive <router-link :to="{ name: 'help' }">Markdown</router-link></span>
           <textarea rows="4" v-model="product.missionStatement" @input="dirty = true" maxlength="320"></textarea>
         </label>
       </div>